Add y/n prompt to upload script
To prevent accidental uploads. Unlike before, snapshot uploads are
prompted now as well, but that can be solved by piping in `yes' in the
cron job.

+++ common/upload.sh
@@ -9,6 +9,12 @@ trap '[ $? -ne 0 ] && echo -e "Missing files; try \`make $type\x27."' EXIT
 xpi=$(basename $(xmlstarlet sel -t -v "//em:updateLink" "$rdf") 2> /dev/null)
 [ -f "$xpi" ] || exit 1
 
+read -p "Upload $xpi? [y/N] "
+if [[ $REPLY != [yY] ]]; then
+    echo "Canceled." >&2
+    exit 0
+fi
+
 case $type in
   snapshot)
     ssh $host "rm -f $dir/snapshot/*.xpi"
